`Development`: Simplify WebsocketMessagingService
