+++ src/main/java/de/tum/in/www1/artemis/web/rest/AssessmentResource.java
@@ -17,12 +17,12 @@ import de.tum.in.www1.artemis.security.Role;
 import de.tum.in.www1.artemis.service.AssessmentService;
 import de.tum.in.www1.artemis.service.AuthorizationCheckService;
 import de.tum.in.www1.artemis.service.ExerciseDateService;
-import de.tum.in.www1.artemis.service.WebsocketMessagingService;
 import de.tum.in.www1.artemis.service.exam.ExamService;
 import de.tum.in.www1.artemis.service.notifications.SingleUserNotificationService;
 import de.tum.in.www1.artemis.web.rest.errors.AccessForbiddenException;
 import de.tum.in.www1.artemis.web.rest.errors.BadRequestAlertException;
 import de.tum.in.www1.artemis.web.rest.errors.EntityNotFoundException;
+import de.tum.in.www1.artemis.web.websocket.ResultWebsocketService;
 
 public abstract class AssessmentResource {
 
@@ -40,7 +40,7 @@ public abstract class AssessmentResource {
 
     protected final ExamService examService;
 
-    protected final WebsocketMessagingService messagingService;
+    protected final ResultWebsocketService resultWebsocketService;
 
     protected final ExampleSubmissionRepository exampleSubmissionRepository;
 
@@ -49,7 +49,7 @@ public abstract class AssessmentResource {
     protected final SingleUserNotificationService singleUserNotificationService;
 
     public AssessmentResource(AuthorizationCheckService authCheckService, UserRepository userRepository, ExerciseRepository exerciseRepository, AssessmentService assessmentService,
-            ResultRepository resultRepository, ExamService examService, WebsocketMessagingService messagingService, ExampleSubmissionRepository exampleSubmissionRepository,
+            ResultRepository resultRepository, ExamService examService, ResultWebsocketService resultWebsocketService, ExampleSubmissionRepository exampleSubmissionRepository,
             SubmissionRepository submissionRepository, SingleUserNotificationService singleUserNotificationService) {
         this.authCheckService = authCheckService;
         this.userRepository = userRepository;
@@ -57,7 +57,7 @@ public abstract class AssessmentResource {
         this.assessmentService = assessmentService;
         this.resultRepository = resultRepository;
         this.examService = examService;
-        this.messagingService = messagingService;
+        this.resultWebsocketService = resultWebsocketService;
         this.exampleSubmissionRepository = exampleSubmissionRepository;
         this.submissionRepository = submissionRepository;
         this.singleUserNotificationService = singleUserNotificationService;
@@ -133,7 +133,7 @@ public abstract class AssessmentResource {
             participation.filterSensitiveInformation();
         }
         if (submit && ExerciseDateService.isAfterAssessmentDueDate(exercise)) {
-            messagingService.broadcastNewResult(result.getParticipation(), result);
+            resultWebsocketService.broadcastNewResult(result.getParticipation(), result);
         }
         return ResponseEntity.ok(result);
     }
